requirements.txt
fastapi==0.115.8
python-dotenv==1.0.1
SQLAlchemy==2.0.37
gunicorn==23.0.0
uvicorn==0.34.0
psycopg2-binary==2.9.10
passlib==1.7.4
bcrypt==4.2.1
pre-commit==4.1.0
python-jose==3.3.0
cryptography==44.0.0
python-multipart==0.0.20